Moved inPeriod function to utils file and filter governing bodies based on id
